@@ -14,32 +14,40 @@ namespace ESFA.DC.LARS.Web.Configuration
     {
         private static int _timeOutInSeconds = 40;
 
-        public static PolicyWrap<HttpResponseMessage> PolicyStrategy => Policy.WrapAsync(RetryPolicy, TimeoutPolicy);
+        //public static PolicyWrap<HttpResponseMessage> PolicyStrategy => Policy.WrapAsync(RetryPolicy, TimeoutPolicy);
 
-        private static TimeoutPolicy<HttpResponseMessage> TimeoutPolicy
-        {
-            get
-            {
-                return Policy.TimeoutAsync<HttpResponseMessage>(_timeOutInSeconds, (context, timeSpan, task) =>
-                {
-                    Debug.WriteLine($"[App|Policy]: Timeout delegate fired after {timeSpan.Seconds} seconds");
-                    return Task.CompletedTask;
-                });
-            }
-        }
+        //private static TimeoutPolicy<HttpResponseMessage> TimeoutPolicy
+        //{
+        //    get
+        //    {
+        //        return Policy.TimeoutAsync<HttpResponseMessage>(_timeOutInSeconds, (context, timeSpan, task) =>
+        //        {
+        //            Debug.WriteLine($"[App|Policy]: Timeout delegate fired after {timeSpan.Seconds} seconds");
+        //            return Task.CompletedTask;
+        //        });
+        //    }
+        //}
 
-        private static RetryPolicy<HttpResponseMessage> RetryPolicy
-        {
-            get
-            {
-                var jitter = new Random();
-                return HttpPolicyExtensions
-                    .HandleTransientHttpError()
-                    .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.NotFound)
-                    .WaitAndRetryAsync(3, retryAttempt =>
-                        TimeSpan.FromSeconds(Math.Pow(2, retryAttempt))
-                        + TimeSpan.FromMilliseconds(jitter.Next(0, 100)));
-            }
-        }
+        //private static RetryPolicy<HttpResponseMessage> RetryPolicy
+        //{
+        //    get
+        //    {
+        //        var jitter = new Random();
+        //        return HttpPolicyExtensions
+        //            .HandleTransientHttpError()
+        //            .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.NotFound)
+        //            .WaitAndRetryAsync(3, retryAttempt =>
+        //                TimeSpan.FromSeconds(Math.Pow(2, retryAttempt))
+        //                + TimeSpan.FromMilliseconds(jitter.Next(0, 100)));
+        //    }
+        //}
+
+        //static IAsyncPolicy<HttpResponseMessage> GetRetryPolicy()
+        //{
+        //    return HttpPolicyExtensions
+        //        .HandleTransientHttpError()
+        //        .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.NotFound)
+        //        .WaitAndRetryAsync(6, retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt)));
+        //}
     }
 }
